requirements/base.txt
Django==3.0.4
django-filter==2.2.0
django-model-utils==4.0.0
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
